Main scraper body, async requests
The scraper is kind of too slow to be used effectively. It would take about an hour to scrape the entire site with the speed it gets now.
However, there are only about 1300 items on the site and 100000 possible items, so with a database of the items that are available on the site, you could scrape that in about 5 minutes or so.

I should think of some performance stuff I can do to this.